+++ app/src/main/java/ca/unb/mobiledev/weatherapp/MinMaxTempData.kt
@@ -1,8 +1,8 @@
 package ca.unb.mobiledev.weatherapp
 
 data class MinMaxTempData(
-    val date: String,    // Assuming you want to display the date
-    val maxTemp: Float,  // Maximum temperature for the day
-    val minTemp: Float,   // Minimum temperature for the day
-    val humidity: Int  // Add humidity field
+    val date: String,    // date
+    val maxTemp: Int,  // Maximum temperature for the day
+    val minTemp: Int,   // Minimum temperature for the day
+    val humidity: Int  // humidity field
 )
